move last helper function to helper.py
