+++ src/cogs/listen.py
@@ -1,6 +1,5 @@
 from discord.ext import commands
 
-from cogs.pods import Pods
 from helpers.helper import Helper
 from services.podgqlservice import PodGQLService
 from utils.subscriptions import subscribe
@@ -33,11 +32,11 @@ class ListenCog(commands.Cog, name="Listen"):
 
     @subscribe(PodGQLService.member_added_listener)
     async def on_project_member_added(self, member_with_project):
-        await Pods.add_or_remove_user_to_pod_tc(self.bot, member_with_project, False)
+        await Helper.add_or_remove_user_to_pod_tc(self.bot, member_with_project, False)
 
     @subscribe(PodGQLService.member_removed_listener)
     async def on_project_member_removed(self, member_with_project):
-        await Pods.add_or_remove_user_to_pod_tc(self.bot, member_with_project, True)
+        await Helper.add_or_remove_user_to_pod_tc(self.bot, member_with_project, True)
 
     @subscribe(PodGQLService.team_edited_listener)
     async def on_project_edited(self, project):
