+++ src/cogs/pods.py
@@ -72,41 +72,11 @@ class Pods(commands.Cog, name="Pods"):
                               "Overflow",
                               MentorFinder.find_a_suitable_mentor(role))
 
-    @commands.command(name='assign_pod')
-    @checks.requires_staff_role()
-    async def assign_pod(self, ctx: commands.Context, team_id, pod_name: str):
-        """Assigns a TEAM to a particular POD"""
-        await Helper.assign_pod_helper(self.bot, team_id, pod_name)
-
     @commands.command(name='assign_pods')
     @checks.requires_staff_role()
     async def assign_pods(self, ctx: commands.Context):
         """Assigns remaining TEAMS to PODS"""
         await Helper.assign_pods_helper(self.bot)
-
-    @staticmethod
-    async def add_or_remove_user_to_pod_tc(bot: discord.ext.commands.Bot, member_with_project, should_be_removed: bool):
-        """Add/remove users to a pod text channel, occurs when someone joins or leaves a team in showcase"""
-        print(member_with_project)
-        discord_id = member_with_project["account"]["discordId"]
-        guild: discord.Guild = await bot.fetch_guild(689213562740277361)
-        showcase_team = await TeamConverter.get_showcase_team_by_id(member_with_project["project"]["id"])
-
-        pod = PodConverter.get_pod_by_id(showcase_team["pod"])
-
-        member: discord.Member = await guild.fetch_member(discord_id)
-        tc = await bot.fetch_channel(pod.tc_id)
-
-        # Occurs when a user left a showcase team and is now being removed from the pod text channel
-        if should_be_removed:
-            await SetPermissions.for_channel_with_discord_member(tc, member, remove=True)
-            await tc.send(embed=GenerateEmbed.user_joins_or_leaves_showcase_team(member_with_project, showcase_team,
-                                                                                 status="leaving"))
-        # Occurs when a user joins a showcase team and is now being added to the pod text channel
-        else:
-            await SetPermissions.for_channel_with_discord_member(tc, member, remove=False)
-            await tc.send(embed=GenerateEmbed.user_joins_or_leaves_showcase_team(member_with_project, showcase_team,
-                                                                                 status="joining"))
 
     @commands.command(name='teams', aliases=['list-teams', 'list_teams', 'listteams'])
     @checks.requires_mentor_role()
